♻️ refactor: usePreventScroll
